Further DI work + cleaning up IEditableSettings interface
